backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.10.0
python-dotenv==1.0.1
httpx==0.27.2
youtube-transcript-api==0.6.2
